chore: Update API base URL in frontend utils

+++ backend/api/cruds/cage.py
@@ -3,6 +3,8 @@ from sqlalchemy.engine import Result
 from sqlalchemy.ext.asyncio import AsyncSession
 
 from api.models.model import Cage
+from datetime import datetime
+import api.schemas.cage as schema
 
 
 async def get_cages(db: AsyncSession):
@@ -21,28 +23,36 @@ async def get_cages(db: AsyncSession):
     return formatted_cages
 
 
-def create_cage(cage_name: str, db: AsyncSession):
-    cage = Cage(cageName=cage_name)
-    db.add(cage)
-    db.commit()
-    db.refresh(cage)
+async def create_cage(cage_create: schema.CageCreate, db: AsyncSession):
+    new_cage = Cage(
+        cageName=cage_create.cageName,
+        createdAt=datetime.now(),
+        updatedAt=datetime.now(),
+    )
+    db.add(new_cage)
+    await db.commit()
+    await db.refresh(new_cage)
+    return new_cage
+
+
+async def update_cage(cage_id: int, new_cage: schema.CageBase, db: AsyncSession):
+    stmt = select(Cage).where(Cage.id == cage_id)
+    result = await db.execute(stmt)
+    cage = result.scalar_one_or_none()
+    cage.cageName = new_cage.cageName
+    cage.updatedAt = datetime.now()
+
+    await db.commit()
+    await db.refresh(cage)
     return cage
 
 
-def update_cage(cage_id: int, cage_name: str, db: AsyncSession):
+async def delete_cage(cage_id: int, db: AsyncSession):
     stmt = select(Cage).where(Cage.id == cage_id)
-    result = db.execute(stmt)
+    result = await db.execute(stmt)
     cage = result.scalars().first()
-    cage.cageName = cage_name
-    db.commit()
-    db.refresh(cage)
-    return cage
-
-
-def delete_cage(cage_id: int, db: AsyncSession):
-    stmt = select(Cage).where(Cage.id == cage_id)
-    result = db.execute(stmt)
-    cage = result.scalars().first()
-    db.delete(cage)
-    db.commit()
+    if cage is None:
+        return None
+    await db.delete(cage)
+    await db.commit()
     return cage
